requirements.txt
mysqlclient
Django
django-cors-headers
django-rest-auth
djangorestframework
